Keep factories files in a manageable size
